fix: test

+++ contracts/test/DummyMarket.sol
@@ -6,9 +6,5 @@ import "../interfaces/IReward.sol";
 import "../interfaces/IFlow.sol";
 
 contract DummyMarket is IMarket {
-    function chargeFee(
-        uint256 beforeLength,
-        uint256 uploadSectors,
-        uint256 paddingSectors
-    ) external {}
+    function chargeFee(uint beforeLength, uint uploadSectors, uint paddingSectors) external {}
 }
